# app/Http/Requests/CampaignFeedbackRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;

class CampaignFeedbackRequest extends FormRequest
{
    public function authorize(): bool
    {
        return true;
    }

    /**
     * @return array<string, mixed>
     */
    public function rules(): array
    {
        return [
            'feedback_qualitativo' => ['required', 'string', 'min:10', 'max:2000'],
        ];
    }
}