# app/Http/Requests/UpdateCampaignRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;

class UpdateCampaignRequest extends FormRequest
{
    public function authorize(): bool
    {
        return true;
    }

    /**
     * @return array<string, mixed>
     */
    public function rules(): array
    {
        return [
            'title' => ['sometimes', 'string', 'max:120'],
            'description' => ['nullable', 'string', 'max:2000'],
            'discount_percent' => ['sometimes', 'numeric', 'min:0.01'],
            'budget_amount' => ['nullable', 'numeric', 'min:0'],
            'radius_meters' => ['sometimes', 'integer', 'min:50'],
            'starts_at' => ['sometimes', 'date'],
            'ends_at' => ['sometimes', 'date', 'after:starts_at'],
        ];
    }
}