# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Postos API",
        "description": "API de gestão de postos (PinPoint): JWT, RBAC, stock e histórico.\r\n\r\n**Rotas públicas (sem JWT):**\r\n- `GET /postos` — listar, paginar (`?page=`) ou detalhe (`?id=`)\r\n- `GET /postos/search` — pesquisa rápida (`?q=`, opcional `limit`, `tipo`, `combustivel`, `status`)\r\n\r\n**Papéis locais (`users.role_id`):** ver `docs/ROLES.md` e `GET /api/roles`.",
        "contact": {
            "email": "suporte@postosapi.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Prefixo das rotas API"
        }
    ],
    "paths": {
...
        "/posts/{post}/campaigns": {
            "get": {
                "tags": [
                    "Gestor",
                    "Campanhas"
                ],
                "summary": "Listar campanhas geolocalizadas (RN-G-007)",
                "operationId": "122649e44dd72f350fed326591bc17fe",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de campanhas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Gestor",
                    "Campanhas"
                ],
                "summary": "Criar campanha (não-combustível, geolocalizada)",
                "operationId": "a7e7bd53dea14f0a8063a8d45768ec54",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCampaignRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Campanha criada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}/campaigns/{campaign}": {
            "patch": {
                "tags": [
                    "Gestor",
                    "Campanhas"
                ],
                "summary": "Actualizar campanha",
                "operationId": "c60ee6419da62d2404118ced3733508a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCampaignRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}/campaigns/{campaign}/performance": {
            "get": {
                "tags": [
                    "Gestor",
                    "Campanhas"
                ],
                "summary": "Métricas de performance (RN-G-007.1)",
                "operationId": "b3a20886954390885526d83fe94d7a18",
                "responses": {
                    "200": {
                        "description": "Métricas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}/campaigns/{campaign}/feedback": {
            "patch": {
                "tags": [
                    "Gestor",
                    "Campanhas"
                ],
                "summary": "Feedback qualitativo da campanha (RN-G-007.1)",
                "operationId": "610702af9603b6deabb5a30972727866",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CampaignFeedbackRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Feedback registado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}/campaigns/{campaign}/pause": {
            "post": {
                "tags": [
                    "Gestor",
                    "Campanhas"
                ],
                "summary": "Pausar campanha activa",
                "operationId": "a1e197b0c61f6e1f651cbc4a073b6966",
                "responses": {
                    "200": {
                        "description": "Pausada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}/campaigns/{campaign}/resume": {
            "post": {
                "tags": [
                    "Gestor",
                    "Campanhas"
                ],
                "summary": "Retomar campanha pausada",
                "operationId": "781ba734f4c7d3af336f220c84643f37",
                "responses": {
                    "200": {
                        "description": "Retomada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}/campaigns/{campaign}/cancel": {
            "post": {
                "tags": [
                    "Gestor",
                    "Campanhas"
                ],
                "summary": "Cancelar campanha",
                "operationId": "95b4da530c9452076a9e77e70d184bab",
                "responses": {
                    "200": {
                        "description": "Cancelada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/posts/{post}/sync": {
            "post": {
                "tags": [
                    "Gestor",
                    "Offline"
                ],
                "summary": "Sincronizar lote offline (RN-G-008)",
                "description": "Processa operações armazenadas localmente com resolução de conflitos.",
                "operationId": "cf3f639b8575068d2dec0196fb3da232",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OfflineSyncBatchRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lote processado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}/sync/batches": {
            "get": {
                "tags": [
                    "Gestor",
                    "Offline"
                ],
                "summary": "Histórico de sincronizações",
                "operationId": "e44e82f5ef7a04614707b6cb3cd7d86e",
                "responses": {
                    "200": {
                        "description": "Lista de lotes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}/sync/batches/{batch}": {
            "get": {
                "tags": [
                    "Gestor",
                    "Offline"
                ],
                "summary": "Detalhe de um lote de sincronização",
                "operationId": "42202b5f638d37010a0140e4ba72a8eb",
                "responses": {
                    "200": {
                        "description": "Detalhe do lote"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreCampaignRequest": {
                "required": [
                    "title",
                    "starts_at",
                    "ends_at"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Super7 - fim de semana"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "discount_percent": {
                        "type": "number",
                        "format": "float",
                        "example": 20
                    },
                    "budget_amount": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "radius_meters": {
                        "type": "integer",
                        "example": 500
                    },
                    "service_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "product_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "starts_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "ends_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "scheduled"
                        ]
                    }
                },
                "type": "object"
            },
            "UpdateCampaignRequest": {
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "discount_percent": {
                        "type": "number",
                        "format": "float"
                    },
                    "budget_amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "radius_meters": {
                        "type": "integer"
                    },
                    "starts_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "ends_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CampaignFeedbackRequest": {
                "required": [
                    "feedback_qualitativo"
                ],
                "properties": {
                    "feedback_qualitativo": {
                        "type": "string",
                        "example": "Promoção gerou muitas vendas na loja."
                    }
                },
                "type": "object"
            },
...
            "OfflineSyncBatchRequest": {
                "required": [
                    "operations"
                ],
                "properties": {
                    "device_id": {
                        "type": "string",
                        "nullable": true
                    },
                    "client_batch_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "operations": {
                        "type": "array",
                        "items": {
                            "required": [
                                "id",
                                "type",
                                "payload",
                                "client_timestamp"
                            ],
                            "properties": {
                                "id": {
                                    "type": "string",
                                    "format": "uuid"
                                },
                                "type": {
                                    "type": "string",
                                    "example": "stock.update"
                                },
                                "payload": {
                                    "type": "object"
                                },
                                "client_timestamp": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT no header Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}